tests: test all notification views
