+++ gsl_notification/templates/gsl_notification/upload_arrete_signe.html
@@ -2,7 +2,7 @@
 
 {% block notification_content %}
     <a class="fr-btn fr-btn--icon-left fr-btn--tertiary fr-icon-arrow-left-s-line fr-mb-4w"
-       href="{% url 'gsl_notification:documents-in-simulation' programmation_projet.id %}">
+       href="{% url 'gsl_notification:documents' programmation_projet.id %}">
         Annuler la génération
     </a>
