+++ gsl_notification/templates/includes/_arrete_form.html
@@ -65,7 +65,7 @@
         <a type="button"
            id="cancel-arrete-form"
            class="fr-btn fr-btn--secondary"
-           href="{% url 'gsl_notification:documents-in-simulation' programmation_projet.id %}">
+           href="{% url 'gsl_notification:documents' programmation_projet.id %}">
             Annuler
         </a>
         <button type="submit" id="submit-arrete-form" class="fr-btn">
